Add files via upload
POC project: Synopsis
Streaming processing of measurement results
There are a number of sensors that measure some physical quantities. The sensor can be connected via UDP protocol and report the measured values. Each sensor can report measurements either at random points in time, or according to its own schedule. The sensor sends the measured values in the form of the following data structure:
typedef struct probe_report_t {
    uint32_t probe_id;
    uint32_t measurement_id;
    int32_t  value;
} probe_report;
where
- probe_id – unique sensor identifier,
- measurement_id – unique measure type identifier,
- value – measured value (signed 32-bit integer).
Different sensors can make the same measurements (device_ids are different, but measurement_ids are the same).
There is a server for processing measurement results. It allows sensors to connect via UDP protocol, receive measured values and store them in memory. For each sensor and measurement, the server averages the values obtained, averaging for each sensor is performed individually, the length of the averaging sequence is set in the server configuration. In addition, the server must remember the start time of the averaging period and the end time of the averaging period for each sensor.
The server can give the last accumulated average values to clients over the TCP protocol, while if the average value has been read by the client, it is immediately erased from the server's memory and instead the server will have to give the string "n/a" to other clients who have requested the value of the same sensor, if by the time of the request the necessary number of measurements for averaging has not yet been accumulated. Together with the average value of the sensor, the server should give the clients information about the time interval for which this average value was received.
There are many user terminals that can connect to the processing server over the TCP protocol. Through the user terminal, a user can request the latest accumulated average for the selected sensor and print it.
It is required to write a software implementation of the measurement results processing server, a simulator program for sensor behavior: the frequency of value generation, device_id and measurement_id must be set for each sensor individually via the command line or in the configuration file, as well as a client program for obtaining measured values from the server.
The sensor simulation program (device_id and measurement_id can be set for example via command line parameters), the server for processing measurement results and the client user terminal applications shall be written for Linux using C and standard C library (POSIX).